List element message type created for BulkExcavation. Experiencing difficulty including necessary header files into ow_regolith
