+++ ow_materials/msg/MaterialConcentration.msg
@@ -1,0 +1,2 @@
+uint8   id          # ID in the material database
+float32 proportion  # concentration of material, sums to unity
