+++ ow_regolith/CMakeLists.txt
@@ -140,6 +140,7 @@ add_dependencies(regolith_node
 target_link_libraries(regolith_node
   ${catkin_LIBRARIES}
   ${GAZEBO_LIBRARIES}
+  Materials
 )
 
 # Gazebo sensor plugin that tracks when regolith collides with the terrain
