+++ ow_regolith/package.xml
@@ -22,7 +22,8 @@
   <!-- <depend>gazebo_ros</depend> -->
 
   <buildtool_depend>catkin</buildtool_depend>
-  
+
+
   <depend>roscpp</depend>
   <depend>gazebo</depend>
   <depend>gazebo_ros</depend>
